leds: Check if the act led trigger file exists.

Former-commit-id: 00a54fbb837d15b441399987aee757cbdbbab40f

+++ overlay/etc/init.d/rockchip.sh
@@ -222,14 +222,15 @@ hwclock --systohc
 /etc/network/ethernet_wol.sh
 
 # set act-led trigger function
-cmdline=$(cat /proc/cmdline)
-storage=`echo $cmdline|awk '{print match($0,"storagemedia=emmc")}'`;
+if [ -e "/sys/class/leds/act-led/trigger" ]; then
+    cmdline=$(cat /proc/cmdline)
+    storage=`echo $cmdline|awk '{print match($0,"storagemedia=emmc")}'`;
 
-if [ $storage -gt 0 ]; then
-    #emmc
-    echo mmc0 > /sys/class/leds/act-led/trigger
-else
-    #sdcard
-    echo mmc1 > /sys/class/leds/act-led/trigger
+    if [ $storage -gt 0 ]; then
+        #emmc
+        echo mmc0 > /sys/class/leds/act-led/trigger
+    else
+        #sdcard
+        echo mmc1 > /sys/class/leds/act-led/trigger
+    fi
 fi
-
